Implementation of administrative functions and validation.

+++ src/lib/actions/data/arquivos.ts
@@ -3,12 +3,22 @@
 import { db } from "@/lib/db";
 import { ServerResponse } from "@/types/ServerResponse";
 import { deleteFileLegacy } from "./legacy/arquivos";
+import { cookies } from "next/headers";
+import { verifyJwt } from "@/lib/jwt";
 
 const isLegacy = process.env.USE_LEGACY_TABLES === "true";
 
 export async function getArchivesById(obraId: number) {}
 
 export async function deleteFile(fileId: number): Promise<ServerResponse> {
+  const adminToken = cookies().get("adminJwt");
+
+  if (!adminToken)
+    return { success: false, error: "Autorização insuficiente." };
+  const isValidAdmin = (await verifyJwt(adminToken.value, true)) !== null;
+  if (!isValidAdmin)
+    return { success: false, error: "Autorização insuficiente." };
+
   try {
     if (isLegacy) return await deleteFileLegacy(fileId);
